# gates do less field.Element copies, mostly. 