--- a/.gitignore
+++ b/.gitignore
@@ -1,1 +1,5 @@
 profiling
+**.prof 
+**.test
+**.txt
+**trace.out