--- a/circuit/gates.go
+++ b/circuit/gates.go
@@ -14,24 +14,24 @@ type Gate interface {
 	// GnarkEval performs the same computation as Eval but on Gnark variables
 	GnarkEval(cs *frontend.ConstraintSystem, vL, vR frontend.Variable) frontend.Variable
 	// Eval returns an evaluation for a unique pair of Eval
-	Eval(res *fr.Element, vL, vR fr.Element)
+	Eval(res, vL, vR *fr.Element)
 	// EvalManyVL returns multiple evaluations with the same vR
-	EvalManyVL(res, vLs []fr.Element, vR fr.Element)
+	EvalManyVL(res, vLs []fr.Element, vR *fr.Element)
 	// EvalManyVR returns multiple evaluations with the same vL
-	EvalManyVR(res []fr.Element, vL fr.Element, vRs []fr.Element)
+	EvalManyVR(res []fr.Element, vL *fr.Element, vRs []fr.Element)
 	// Degrees returns the degrees of the gate relatively to HL, HR, HPrime
 	Degrees() (degHL, degHR, degHPrime int)
 }
 
 // EvaluateCombinator evaluate eq * \sum_i{ statics_i * gates_i(vL, vR) }
-func EvaluateCombinator(vL, vR, eq fr.Element, gates []Gate, statics []fr.Element) fr.Element {
+func EvaluateCombinator(vL, vR, eq *fr.Element, gates []Gate, statics []fr.Element) fr.Element {
 	var tmp, res fr.Element
 	for i := range gates {
 		gates[i].Eval(&tmp, vL, vR)
 		tmp.Mul(&statics[i], &tmp)
 		res.Add(&res, &tmp)
 	}
-	res.Mul(&res, &eq)
+	res.Mul(&res, eq)
 	return res
 }
 
@@ -42,8 +42,8 @@ type AddGate struct{}
 func (a AddGate) ID() string { return "AddGate" }
 
 // Eval return the result vL + vR
-func (a AddGate) Eval(res *fr.Element, vL, vR fr.Element) {
-	res.Add(&vL, &vR)
+func (a AddGate) Eval(res, vL, vR *fr.Element) {
+	res.Add(vL, vR)
 }
 
 // GnarkEval compute the gate on a gnark circuit
@@ -54,9 +54,9 @@ func (a AddGate) GnarkEval(cs *frontend.ConstraintSystem, vL, vR frontend.Variab
 
 // EvalManyVR performs an element-wise addition of many vRs values by one vL value
 // res must be initialized with the same size as vRs
-func (a AddGate) EvalManyVR(res []fr.Element, vL fr.Element, vRs []fr.Element) {
+func (a AddGate) EvalManyVR(res []fr.Element, vL *fr.Element, vRs []fr.Element) {
 	for i, vR := range vRs {
-		res[i].Add(&vL, &vR)
+		res[i].Add(vL, &vR)
 	}
 }
 
@@ -79,8 +79,8 @@ type MulGate struct{}
 func (m MulGate) ID() string { return "MulGate" }
 
 // Eval returns vL * vR
-func (m MulGate) Eval(res *fr.Element, vL, vR fr.Element) {
-	res.Mul(&vL, &vR)
+func (m MulGate) Eval(res, vL, vR *fr.Element) {
+	res.Mul(vL, vR)
 }
 
 // GnarkEval performs the gate operation on gnark variables
@@ -89,16 +89,16 @@ func (m MulGate) GnarkEval(cs *frontend.ConstraintSystem, vL, vR frontend.Variab
 }
 
 // EvalManyVR performs an element-wise multiplication of many vRs values by one vL value
-func (m MulGate) EvalManyVR(res []fr.Element, vL fr.Element, vRs []fr.Element) {
+func (m MulGate) EvalManyVR(res []fr.Element, vL *fr.Element, vRs []fr.Element) {
 	for i, vR := range vRs {
-		res[i].Mul(&vL, &vR)
+		res[i].Mul(vL, &vR)
 	}
 }
 
 // EvalManyVL performs an element-wise multiplication of many vLs values by one vR value
-func (m MulGate) EvalManyVL(res, vLs []fr.Element, vR fr.Element) {
+func (m MulGate) EvalManyVL(res, vLs []fr.Element, vR *fr.Element) {
 	for i, vL := range vLs {
-		res[i].Mul(&vL, &vR)
+		res[i].Mul(&vL, vR)
 	}
 }
 
@@ -114,8 +114,9 @@ type CopyGate struct{}
 func (c CopyGate) ID() string { return "CopyGate" }
 
 // Eval returns vL
-func (c CopyGate) Eval(res *fr.Element, vL, vR fr.Element) {
-	*res = vL
+func (c CopyGate) Eval(res, vL, vR *fr.Element) {
+	res.Set(vL)
+	// *res = vL
 }
 
 // GnarkEval performs the copy on gnark variable
@@ -124,14 +125,14 @@ func (c CopyGate) GnarkEval(cs *frontend.ConstraintSystem, vL, vR frontend.Varia
 }
 
 // EvalManyVR performs an element-wise copy of vL for many vRs. (ignoring the values of the vRs)
-func (c CopyGate) EvalManyVR(res []fr.Element, vL fr.Element, vRs []fr.Element) {
+func (c CopyGate) EvalManyVR(res []fr.Element, vL *fr.Element, vRs []fr.Element) {
 	for i := range vRs {
-		res[i] = vL
+		res[i].Set(vL)
 	}
 }
 
 // EvalManyVL performs an element-wise copy of many vLs values
-func (c CopyGate) EvalManyVL(res, vLs []fr.Element, vR fr.Element) {
+func (c CopyGate) EvalManyVL(res, vLs []fr.Element, vR *fr.Element) {
 	for i := range vLs {
 		res[i] = vLs[i]
 	}
@@ -148,18 +149,18 @@ type CipherGate struct {
 }
 
 // NewCipherGate construct a new cipher gate given an ark
-func NewCipherGate(ark fr.Element) CipherGate {
-	return CipherGate{Ark: ark}
+func NewCipherGate(ark fr.Element) *CipherGate {
+	return &CipherGate{Ark: ark}
 }
 
 // ID returns the id of the cipher gate and print the ark as well
-func (c CipherGate) ID() string { return fmt.Sprintf("CipherGate-%v", c.Ark.String()) }
+func (c *CipherGate) ID() string { return fmt.Sprintf("CipherGate-%v", c.Ark.String()) }
 
 // Eval returns vL + (vR + c)^7
-func (c CipherGate) Eval(res *fr.Element, vL, vR fr.Element) {
+func (c *CipherGate) Eval(res, vL, vR *fr.Element) {
 	// tmp = vR + Ark
-	tmp := vR
-	tmp.Add(&tmp, &c.Ark)
+	var tmp fr.Element
+	tmp.Add(vR, &c.Ark)
 	// res = tmp^7
 	*res = tmp
 	res.Square(res)
@@ -167,11 +168,11 @@ func (c CipherGate) Eval(res *fr.Element, vL, vR fr.Element) {
 	res.Square(res)
 	res.Mul(res, &tmp)
 	// Then add vL
-	res.Add(res, &vL)
+	res.Add(res, vL)
 }
 
 // GnarkEval performs the cipher operation on gnark variables
-func (c CipherGate) GnarkEval(cs *frontend.ConstraintSystem, vL, vR frontend.Variable) frontend.Variable {
+func (c *CipherGate) GnarkEval(cs *frontend.ConstraintSystem, vL, vR frontend.Variable) frontend.Variable {
 	tmp := cs.Add(vR, cs.Constant(c.Ark))
 	cipher := cs.Mul(tmp, tmp)
 	cipher = cs.Mul(cipher, tmp)
@@ -182,18 +183,18 @@ func (c CipherGate) GnarkEval(cs *frontend.ConstraintSystem, vL, vR frontend.Var
 
 // EvalManyVR performs cipher evaluations of many vRs values by one vL value
 // Nothing special to do here
-func (c CipherGate) EvalManyVR(res []fr.Element, vL fr.Element, vRs []fr.Element) {
+func (c *CipherGate) EvalManyVR(res []fr.Element, vL *fr.Element, vRs []fr.Element) {
 	for i, vR := range vRs {
-		c.Eval(&res[i], vL, vR)
+		c.Eval(&res[i], vL, &vR)
 	}
 }
 
 // EvalManyVL performs an element-wise cipher of many vLs values by one vR
 // This one is optimized to only do the vL exponentiation once
-func (c CipherGate) EvalManyVL(res, vLs []fr.Element, vR fr.Element) {
+func (c *CipherGate) EvalManyVL(res, vLs []fr.Element, vR *fr.Element) {
 	// tmp = vR + Ark
-	tmp := vR
-	tmp.Add(&tmp, &c.Ark)
+	var tmp fr.Element
+	tmp.Add(vR, &c.Ark)
 	// right = tmp^7
 	right := tmp
 	right.Square(&right)
@@ -207,6 +208,6 @@ func (c CipherGate) EvalManyVL(res, vLs []fr.Element, vR fr.Element) {
 }
 
 // Degrees returns the degrees of the gate on hL, hR and hPrime
-func (c CipherGate) Degrees() (degHL, degHR, degHPrime int) {
+func (c *CipherGate) Degrees() (degHL, degHR, degHPrime int) {
 	return 1, 7, 7
 }