for 

```
BN_GKR=17 NCHUNKS_GKR=8 go test -run=NONE -bench=^BenchmarkMimcGKRProver$ -count=1   | tee main.txt
```

```
benchmark                                old ns/op       new ns/op      delta
BenchmarkMimcGKRProver/bN=17-nCore-8     10628550024     9369816014     -11.84%
```